Form a new component for creating a question
